+++ NEWS.md
@@ -5,6 +5,10 @@
 - In the `ms_drive_item$load_dataframe()` method, pass the `...` argument to `read_delim`.
 - Add the ability to load Excel files (with extension .xls or .xlsx) to the `ms_drive_item$load_dataframe()` method. This requires the readxl package to be installed.
 
+## Teams
+
+- Add the ability to create shared channels (#174).
+
 ## Planner
 
 - Fix a bug in the `ms_plan$get_details()` method.